+++ android/src/main/java/com/bugsnag/BugsnagReactNative.java
@@ -9,6 +9,7 @@ import com.bugsnag.android.Bugsnag;
 import com.bugsnag.android.Callback;
 import com.bugsnag.android.Client;
 import com.bugsnag.android.Configuration;
+import com.bugsnag.android.InternalHooks;
 import com.bugsnag.android.JsonStream;
 import com.bugsnag.android.MetaData;
 import com.bugsnag.android.Report;
@@ -104,6 +105,7 @@ public class BugsnagReactNative extends ReactContextBaseJavaModule {
       libraryVersion = options.getString("version");
       bugsnagAndroidVersion = client.getClass().getPackage().getSpecificationVersion();
       configureRuntimeOptions(client, options);
+      InternalHooks.configureClient(client);
 
       logger.info(String.format("Initialized Bugsnag React Native %s/Android %s",
                   libraryVersion,